+++ out/templates/setup.template.js
@@ -179,7 +179,10 @@ function getMainFileContent(projectName) {
 import 'package:${projectName}/utils/theme.dart';
 import 'package:${projectName}/utils/route_handling.dart';
 
-  void main() {
+import 'dependency_injection.dart' as di;
+
+  void main() async{
+    await di.setup();
     runApp(const MyApp());
   }