+++ src/commands/index.ts
@@ -1,2 +1,4 @@
 export * from "./setup.command";
 export * from "./new-screen.command";
+export * from "./new-repository.command";
+export * from "./new-service.commands";