+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to the "flutter-project-template" extension will be document
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.3.0] - 2023-01-15
+
+### Added
+
+- Template for base repositories
+- Template for firestore repositories
+- Template for base services
+
+### Fixed
+
+- Setup template now includes the import of the dependency_injection file
+
+## [0.2.1] - 2023-01-05
+
+### Fixed
+
+- fpsimpleroute snippet not found
+
 ## [0.2.0] - 2023-01-04
 
 ### Added